+++ lib/pragma_design_system.dart
@@ -25,3 +25,4 @@ export 'src/widgets/pragma_dropdown_list_widget.dart';
 export 'src/widgets/pragma_dropdown_option.dart';
 export 'src/widgets/pragma_dropdown_widget.dart';
 export 'src/widgets/pragma_icon_button.dart';
+export 'src/widgets/pragma_icon_button_widget.dart';